requirements.txt
PyYAML~=6.0.2
mysql~=0.0.3
mysql-connector-python~=9.2.0